Add a simple HTML DSL
A different approach than that taken by Hiccup.jl.

All nodes are the same type (`Node`) and no symbolic `tag` type
parameter. Each node's children are all also `Node`s and attributes are
a `Vector` rather than a `Dict` with concrete type parameters (`Symbol`
and `String`).

Also has a bit of syntax sugar:

    div[".class#id"](
        ul(map(li, split("foo bar baz"))),
        img[:src => "path/to/file.jpg"]
    )

No tags are exported at all. They can just be defined when needed using

    const div, ul, li, img = map(Tag, (:div, :ul, :li, :img))

We can add a macro to do that as well if we do use this.